Not working Network Initialization
